requirements.txt
anthropic>=0.39.0
msal>=1.28.0
requests>=2.31.0
python-docx>=1.1.2
pypdf>=4.3.0
pdfplumber>=0.11.0
pillow>=10.4.0
pydantic>=2.8.0
python-dotenv>=1.0.1
tenacity>=8.3.0
